Use DECIMAL(20, 16) instead of DOUBLE

+++ MySQL_Advanced/10-div.sql
@@ -7,7 +7,7 @@
     -- And returns a / b or 0 if b == 0
 DELIMITER $$
 CREATE FUNCTION SafeDiv (a INT, b INT)
-    RETURNS DOUBLE DETERMINISTIC
+    RETURNS DECIMAL(20, 16) DETERMINISTIC
     BEGIN
         RETURN IF(b=0, 0, a/b);
     END
